ByteNCrunch/requirements.txt

mysql-connector-python==8.0.33
python-dotenv==1.0.0
python-telegram-bot==13.0

